Switch to u64 Balance

+++ common/src/lib.rs
@@ -85,7 +85,7 @@ pub const DATA_PULL_BYTES_BEFORE_LEN: u16 = 16; // How many bytes before the pul
 // python3 -c "from datetime import datetime; print(int(datetime.strptime('2024-01-01 00:00:00', '%Y-%m-%d %H:%M:%S').timestamp()), '* 1_000_000_000')"
 pub const FIRST_BLOCK_TIMESTAMP_NS: u64 = 1704063600 * 1_000_000_000;
 
-pub type Balance = u128;
+pub type Balance = u64;
 
 pub fn nat_to_balance(nat: &Nat) -> Balance {
     nat.0
